Modify evaluation of signal power
In this commit, I'm now vvaluating the signal power on the actual Doppler bins instead of finding the max of the Doppler FFT signal power. Also fixed the seed to 3 else I'm not getting the same input and output SNRs till beofre phase noise kicks in. Need to debug and fix this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>