Fix bug in generating target range
In this commit, I have fixed a bug in generating the target range. Due to this bug, sometimes, the target range was coming out as negative. Since we throw out the negative frequencies in range, the signal was getting lost and some random incorrect range bin was getting sampled and this was leading to some weird results. This was the reason, the output was very dependent on the seed programmed. For any random run, sometimes, it was generating a negative range. Now, I have fixed the target to range to 1*rangeRes and made the max range bins post performing RFFT to just 3 bins. This is because, all other range bins are just noise and I dont need to process the other range bins. The compression analysis needs to be done only on the target range bin and hence we can reduce the compute by limiting the processing to just the target range bin. But I have given a further processing of 2 range bins.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>