Improve spur levels in Conti compression scheme
Conti scheme suffers from larger precision loss at higher SNRs. So, in general, at higher SNRs, the quantization loss is higher and if we simply pad zeros after the mantissa bits, then we incur a larger loss and this error can get periodic resulting in spurs. In their patent, Conti suggests a way to reduce these spurs. Instead of simply padding zeros post the mantissa bits, we replace it with 1 followed by zeros. This way, we incur a lower loss on average since we take a middle ground. The bits post the mantissa could vary from all 0's (lowest value) to all 1's (highest value). So, we replace it with 1 followed by zeros which is like the midpoint. This ensures we reduce the error on "average" and reduces the spurs levels significantly. Conti suggests adding a random number ranging from 0 to mantissa LSB. But for now, I have done a soft rounding.

I have made one more change in this commit. For checking the compression analysis, I was dialling 0 Doppler. But this was somehow not showing up all issues like spurs, etc. So, now, I have given a small velocity to the targets in the simulations.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>