Improve compute time of Huffman encoding
In this commit, I have drastically improved the compute time of the data encoding part of the Hufman compression scheme. Previously, I was obtaining the code word for each symbol and appending them one after the other to create a long binary string of 0s and 1s for all the symbols. Then, I was taking 8 bits at a time from this very long string and creating a byte array.
Now, what I do is, keep appending the code words(Huffman codes of each symbol) into a binary string till it reaches a length of 8 i.e 1 byte, then I pack it to a byte array then itself and keep appending into this byte array as I move through the data symbols. If the size of the binary string is greater than 8, then I pick first 8 bits and pack it to a byte array and the remaining bits are used to concatenate to the code word of the next data symbol. So, this way, instead of creating a long binary string and then packing 8 bits into a byte array, now, as and when I get 8 length binary string, I pack into a byte array there itself and keep appending these byte arrays as I move through the data symbols. This optimization has greatly improved the encoding time! Also, bitmatched the uncompressed and decompressed files with this new scheme.

I have also added the compute times for the encoding and decoding part.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>