Replace Decoding for loop with inbuilt list functions
In the Huffman decoding part, I was looping(for loop) over all the dictionary items (key value pairs) to find the key/symbol for the binary string in the encoded data. Now, I have replaced the for loop with inbuilt list functions which directly gives the associated key/symbol for the corresponding value/binary string.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>