Optimize the Huffman decoding process
In this commit, I have optimized the Huffman decoding process. Now, the decoding time has reduced by almost 50% especially for large data files. Previously, I was first searching whether the binary string/codeword was present in the dictionary or not using the inbuilt 'in' function for dictionary. Only if it present in the dictionary, then I was finding out the symbol corresponding to the codeword. Now this operation of trying to find whether the codeword is present in the dictionary has a O(n) computation and this scales to very large values when the length of the dictionary is very large. Also, if the binary string is not found in the dictionary, then another binary character from the encoded data is appened to the binary string increasing the length of the binary string by 1 and again performing the search operation to check if the new binary string is present in the dictionary values. So this search operation which is already compute heavy is also in while loop! This is a very compute heavy operation. I have worked out a way to reduce this compute. I modify the dictionary of symbol-codeword pairs to 3 list arrays which are all sorted by the length of the codewords. One list contains only the symbols. The second list contains the binary string/codewords. The third list contains the length of each codeword. And this list of length of code words is sorted in the increasing order. This same index is used to sort the other 2 lists. This is done in the method named 'codeWordStructure' in the huffman coding class. In a way, I'm converting the dictionary to a structure with fields as symbols, codewords, length of code words. This way of organizing the Huffman table(sorting in the order of increasing length of the codewords) helps to reduce the compute and here's how.

The search process of finding the symbol corresponding to the codeword begins by taking 1 character of the encoded binary data. If symbol for this character is not found, then the next binary character of the encoded data is appended to the previous string and this new 2 length string is searched for in the dictionary of symbol-codeword pairs. So, each time, the length of the search string is increased by 1 if the previous string is not found. Now, since the length of the search string is progressively increasing by 1, we need to search in the list of codwords, only those code words whose length is equal to the length of the binary string. For example, say the length of the binary string is 3 i.e then we need to search for codewords whose length is 3 and hence we dont have to search over all the other codewords. If the 3 length binary string is not a code word, then we append the next binary character in the encoded data to make the binary string of length 4. Now, also, we need to search for code words whose length is 4 only. Also, when starting the search for binary string of length 4, we can start off from where we stopped for codewords of length 3. This way, by arranging the Huffman decoding LUT in the increasing order of the lenght of the codewords, we can reduce the number of search operations especially for codewords which occur frequently and whose length is small. We just have to search over the first few entries of the Huffman LUT and not over all the entries of the LUT. The major advantages of this optimization are summarized as follows:
1. Reduce the number of search operations by searching over the code words having the same length as the binary string
2. If code word for binary string of length N not found, then start search for the binary string of length N+1 from the index of the last binary string of length N.
3. Arrange the Huffman decoding LUT as 3 lists instead of one dictionary. And these lists should be arranged in the increasing order of lengths of codewors and this index should be applied to the other 2 arrays to maintain the symbol-codeword mapping.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>